Components referenced within geppetto moved out of dev folder
popup/modal components that are tightly bound to the core application
must be included in the core application.
