+++ src/main/webapp/js/components/bower.json
@@ -10,7 +10,6 @@
         "cameracontrols": "https://github.com/mlolson/geppetto-components/raw/master/cameracontrols.zip",
         "help": "https://github.com/mlolson/geppetto-components/raw/master/help.zip",
         "tutorial": "https://github.com/mlolson/geppetto-components/raw/master/tutorial.zip",
-        "popups": "https://github.com/mlolson/geppetto-components/raw/master/popups.zip",
         "bootstrap": "https://github.com/mlolson/geppetto-components/raw/master/bootstrap.zip"
     }
 }
