--- src/main/webapp/js/components/dev/bootstrap/.bower.json
@@ -1,30 +1,0 @@
-{
-  "name": "bootstrap",
-  "version": "0.0.1",
-  "authors": [
-    "Matt Olson <mattlolson@gmail.com>"
-  ],
-  "description": "bootstrap react components",
-  "moduleType": [
-    "amd"
-  ],
-  "license": "MIT",
-  "private": true,
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    ".",
-    "test",
-    "tests"
-  ],
-  "_cacheHeaders": {
-    "ETag": "\"71f5110851fe034920424fda75eefb6fb8b38153\"",
-    "Content-Length": "2393",
-    "Content-Type": "application/zip"
-  },
-  "_release": "e-tag:71f511085",
-  "_source": "https://github.com/mlolson/geppetto-components/raw/master/bootstrap.zip",
-  "_target": "*",
-  "_originalSource": "https://github.com/mlolson/geppetto-components/raw/master/bootstrap.zip"
-}
