--- src/main/webapp/js/components/dev/bootstrap/bower.json
@@ -1,21 +1,0 @@
-{
-  "name": "bootstrap",
-  "version": "0.0.1",
-  "authors": [
-    "Matt Olson <mattlolson@gmail.com>"
-  ],
-  "description": "bootstrap react components",
-  "moduleType": [
-    "amd"
-  ],
-  "license": "MIT",
-  "private": true,
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    ".",
-    "test",
-    "tests"
-  ]
-}
