+++ src/main/webapp/js/components/dev/help/HelpModal.js
@@ -1,12 +1,12 @@
 define(function (require) {
 
     var React = require('react'),
-        Button = require('../bootstrap/button'),
+        Button = require('mixins/bootstrap/button'),
         GEPPETTO = require('geppetto');
 
     return React.createClass({
         mixins: [
-            require('jsx!../bootstrap/modal')
+            require('jsx!mixins/bootstrap/modal')
         ],
 
         startTutorial: function() {
