--- src/main/webapp/js/components/dev/popups/bower.json
@@ -1,21 +1,0 @@
-{
-  "name": "popups",
-  "version": "0.0.1",
-  "authors": [
-    "Matt Olson <mattlolson@gmail.com>"
-  ],
-  "description": "Modals for geppetto",
-  "moduleType": [
-    "amd"
-  ],
-  "license": "MIT",
-  "private": true,
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    ".",
-    "test",
-    "tests"
-  ]
-}
