+++ src/main/webapp/js/components/install_components.py
@@ -14,7 +14,7 @@ for dependency in config['dependencies']:
     componentConfig = json.loads(open(os.path.join(os.path.dirname(__file__), 'dist/'+dependency+'/bower.json')).read())
 
     if 'main' in componentConfig:
-        componentsfile.write("require('jsx!./dist/"+dependency+"/"+componentConfig['main']+"');\n")
+        componentsfile.write("require('jsx!./dev/"+dependency+"/"+componentConfig['main']+"');\n")
 
 componentsfile.write('});');
 componentsfile.close();
