+++ src/main/webapp/js/components/popups/ErrorModal.js
@@ -37,12 +37,12 @@
  */
 define(function (require) {
 
-    var React = require('react')
+    var React = require('react'),
         $ = require('jquery');
 
     return React.createClass({
         mixins: [
-            require('jsx!../bootstrap/modal')
+            require('jsx!mixins/bootstrap/modal')
         ],
         
         getDefaultProps: function() {
