+++ src/main/webapp/js/geppetto-objects/G.js
@@ -44,7 +44,7 @@ define(function(require) {
 		var debugMode = false;
 		var $ = require('jquery'),
 		React = require('react'),
-        ClipboardModal = require('jsx!components/dev/popups/ClipboardModal');
+        ClipboardModal = require('jsx!components/popups/ClipboardModal');
 
 		/**
 		 * @exports geppetto-objects/G
