+++ src/main/webapp/js/mixins/bootstrap/modal.js
